requirements.txt
FastAPI==0.104.1
gunicorn==21.2.0
uvicorn==0.24.0.post1
python-decouple==3.8
pytest==7.4.3
httpx==0.25.2
google-auth==2.25.2
google-cloud-secret-manager==2.17.0
google-cloud-bigquery==3.13.0